refactor: implement WebGPU model and controller for enhanced graphics rendering
