+++ src/lib/webgpu/index.js
@@ -1,0 +1,2 @@
+export { WebGPUModel } from './WebGPUModel.js'
+export { WebGPUController } from './WebGPUController.js'
